⚡️(lti) improve bundle size polyfill
Import only the polyfill needed for the browser.
Replace `require` by `import` to be able to use tree shaking.
