+++ src/frontend/apps/lti_site/package.json
@@ -104,6 +104,7 @@
     "@codemirror/state": "6.2.0",
     "@codemirror/view": "6.7.2",
     "@converse/headless": "8.0.1",
+    "@formatjs/intl-pluralrules": "5.1.8",
     "@formatjs/intl-relativetimeformat": "11.1.8",
     "@mdx-js/mdx2": "npm:@mdx-js/mdx@2.2.1",
     "@sentry/browser": "7.29.0",
@@ -119,7 +120,6 @@
     "highlight.js": "11.7.0",
     "iframe-resizer": "4.3.2",
     "intl": "1.2.5",
-    "intl-pluralrules": "1.3.1",
     "jwt-decode": "3.1.2",
     "katex": "0.16.4",
     "linkifyjs": "4.1.0",
